backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
azure-ai-inference==1.0.0b4
mistralai==1.0.0
pydantic==2.6.4
python-dotenv==1.0.1
duckduckgo-search==5.2.0
pandas==2.2.1
xgboost==2.0.3
shap==0.45.0
reportlab==4.1.0
python-multipart==0.0.9
